Refactor waist-to-hips measurement
